Cmake build: Omit use -f '-Wno-deprecated-copy' on compilers which do not support this flag.
Patch from P. De Visschere

Citing Patrick: As is stands the patch unsets the test-variable
from the cache and the test is performed for each cmake-run.
Otherwise after running cmake once the check is never run again, until
one throws away the cmakecache.txt, which I try to avoid.
I suppose that after time this check will be forgotten and if the issue
is solved (by an upgrade of the Apple clang compiler) this will pass
unnoticed.