Noweb.lyx: fix a few mistakes from old lyx2lyx
In fb034884 I made some manual changes to documents that weren't
correctly converted by lyx2lyx (from ERT to Chunk insets) but I left
a "@" inside a chunk, which ended it prematurely and caused
incorrect output.

This current commit also cleans up a few other things in the
document.

Thanks to Kornel for catching this.