lyx2lyx: prefix warning messages with "lyx2lyx"
Both lyx2lyx and LyX write output to the terminal, and it's helpful
to know which messages are coming from lyx2lyx. For example, before
this commit if we opened a document in LyX that has a newer file format than
lyx2lyx can deal with, we see the following output in the terminal:

  Warning: 619: Format not supported.
  Warning: Quitting.
  Error: Document format failure

The first two lines (the warnings) are output from lyx2lyx, and the
third is output by LyX. This output was particularly confusing
because I first thought "LyX tries to quit if the document is too
new?", but in fact LyX still stays open; it just doesn't open the
document. After this change, the output is now the following:

  lyx2lyx warning: 619: Format not supported.
  lyx2lyx warning: Quitting.
  Error: Document format failure