Fix potential crash with 32bit ucs4 codepoints
QChar codepoints are 16 bit, so it is not wise to pass blindly a
char_type value to the QChar constructor.

Use the already defined isSpace() helper tht takes care of this issue.

Related to #12519, which is only known to bite with Qt6 for now; this
bug is thus theoretical in 2.3.