Update format of docs using LyX binary
Especially after the change to use semantic linefeeds (7b23c76b),
the diffs are large and it's hard to figure out what diff is the
result of the linefeed change and which diff is the result of an
edit.

By updating the docs, it will make the edits easier to understand
from the diff.

This commit used the LyX binary to write the new .lyx files since
lyx2lyx does not apply semantic linefeeds.

I used the following command:

  ./development/tools/updatedocs.py [path to 'lyx' binary]

A few documents were not updated (e.g., the command sequence used
failed because a dialog about a missing dependency was shown).

I did not update Additional.lyx for any language since it is
undergoing changes. I also didn't change anything in doc/fr since JP
is working on some of those. I also didn't change doc/ru since it
appears Yuriy already updated those.