upgrade boost to 1.75.0
Boost is updated to a more recent version to avoid many compilation
warnings. However, boost 1.76 introduces a brand new regex library.
Therefore, it was chosen to use the last version before that to
minimize changes.

The part that adapts our source to the new boost requirements is
backported from 09130d7a622.