Fix bug #12711
Add a few missing autocorrect combinations and allow cycling
through arrows.

The way autocorrect is implemented it is not possible to map
--> and ==> to \longleftarrow and \Longleftarrow, respectively.
One has to type ->- and =>= to that effect, which may be
counterintuitive.

This commit allows typing -> or => followed by * to cycle through
all possible arrows of the same type.